requirements.txt
# Gonka.ai Infrastructure Dependencies
#
# Install: pip install -r infrastructure/requirements.txt

# Gateway
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pyyaml>=6.0

# Model Serving (install separately on GPU nodes)
# vllm>=0.15.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24.0
locust>=2.30.0
